Move login logic to common utils
